Re-remove failing test, added unit test for user deletion in DB
